git: added difftool
